Add files via upload
"A to-do list is a tool used to organize tasks and activities, helping individuals stay focused and productive. It serves as a visual reminder of pending tasks, often allowing users to prioritize, schedule, and track progress efficiently. Whether on paper or digital format, a to-do list aids in time management and goal achievement by providing a clear road-map of what needs to be accomplished."